Completed support for the CAN-bus; it's ready for debugging. Some important details to note include:
- Created a definition for extended-IDs to facilitate waiting for responses. For more details, see the comments in can_messenger.rs.
- Implemented a future method for waiting for responses.
- Moved the parts of mc_main that should be shared to mc/mod.rs.
- Modified the system_executor to support both UART and CAN-bus simultaneously.

+++ pizzaro/src/common/consts.rs
@@ -12,6 +12,6 @@ pub const BELT_OFF_SPEED: i32 = 0;
 pub const LINEAR_BULL_MAX_PRESSURE: u32 = 1000;
 
 pub const CANBUS_FREQUENCY: u32 = 10000;
-pub const MC_CAN_ID: u16 = 1;
-pub const MMD_CAN_ID: u16 = 2;
-pub const HPD_CAN_ID: u16 = 3;
+pub const MC_CAN_ID: u8 = 1;
+pub const MMD_CAN_ID: u8 = 2;
+pub const HPD_CAN_ID: u8 = 3;
